refactor node management UI, update sidebar icon, enhance login form state handling, and modify byte conversion functions

+++ server/utils/bytecovertor.go
@@ -1,13 +1,13 @@
 package utils
 
-func BytesToKB(bytes int64) float64 {
+func BytesToKB(bytes float64) float64 {
 	return float64(bytes) / 1024
 }
 
-func BytesToMB(bytes int64) float64 {
+func BytesToMB(bytes float64) float64 {
 	return float64(bytes) / (1024 * 1024)
 }
 
-func BytesToGB(bytes int64) float64 {
+func BytesToGB(bytes float64) float64 {
 	return float64(bytes) / (1024 * 1024 * 1024)
 }
